test foreign key off
